+++ rclcpp_action/src/server.cpp
@@ -269,7 +269,7 @@ ServerBase::take_data()
 }
 
 std::shared_ptr<void>
-ServerBase::take_data_by_entity_id(int id)
+ServerBase::take_data_by_entity_id(size_t id)
 {
   // Mark as ready the entity from which we want to take data
   switch (static_cast<EntityType>(id)) {
